feat: implement ShoppingListSummaryCard component and refactor list card components
