+++ client/src/pages/home/components/active-lists-section/components/shopping-list-card/index.tsx
@@ -1,15 +1,5 @@
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "@components/dropdown-menu";
-import { Pressable, View } from "react-native";
-import { Card, CardContent } from "@components/card";
-import { Icon } from "@components/icon";
-import { Text } from "@components/text";
+import { ShoppingListSummaryCard } from "@components/shopping-list-summary-card";
 import { ShoppingList } from "@services/shopping-lists/types";
-import { formatUpdatedAt } from "../../../../utils";
 
 interface ShoppingListCardProps {
   isDeleting: boolean;
@@ -20,59 +10,12 @@ interface ShoppingListCardProps {
 
 export const ShoppingListCard = ({ isDeleting, list, onDelete, onOpen }: ShoppingListCardProps) => {
   return (
-    <Pressable onPress={() => onOpen(list.id)} className="flex-1">
-      <Card className="min-h-[150px] justify-between border-0 bg-white py-5 shadow-sm">
-        <CardContent className="flex-1 justify-between gap-5">
-          <View className="flex-row items-start justify-between gap-3">
-            <Pressable className="min-w-0 flex-1" onPress={() => onOpen(list.id)}>
-              <Text numberOfLines={1} className="text-lg font-bold text-foreground">
-                {list.name}
-              </Text>
-              <Text className="font-questrial text-sm text-muted-foreground">
-                {formatUpdatedAt(list.updatedAt)}
-              </Text>
-            </Pressable>
-
-            <DropdownMenu>
-              <DropdownMenuTrigger className="size-8 items-center justify-center rounded-full bg-primary/10">
-                <Icon name="Ellipsis" size={16} className="text-primary" />
-              </DropdownMenuTrigger>
-              <DropdownMenuContent align="end" side="bottom" sideOffset={8}>
-                <DropdownMenuItem
-                  disabled={isDeleting}
-                  variant="destructive"
-                  onPress={() => onDelete(list)}
-                >
-                  <Icon name="Trash2" size={16} className="text-destructive" />
-                  <Text>{isDeleting ? "Excluindo..." : "Excluir lista"}</Text>
-                </DropdownMenuItem>
-              </DropdownMenuContent>
-            </DropdownMenu>
-          </View>
-
-          <View className="flex-row items-center justify-between">
-            <View className="flex-row items-center">
-              {list.items.slice(0, 3).map(item => (
-                <View
-                  key={item.id}
-                  className="-mr-2 size-8 items-center justify-center rounded-full border-2 border-white bg-accent"
-                >
-                  <Icon name="Package" size={14} className="text-muted-foreground" />
-                </View>
-              ))}
-              {list.items.length === 0 ? (
-                <Text className="font-questrial text-sm text-muted-foreground">Sem itens</Text>
-              ) : null}
-            </View>
-
-            <View className="rounded-full bg-primary/10 px-2 py-1">
-              <Text className="font-questrial text-[10px] uppercase text-primary">
-                {list.items.length} {list.items.length === 1 ? "item" : "itens"}
-              </Text>
-            </View>
-          </View>
-        </CardContent>
-      </Card>
-    </Pressable>
+    <ShoppingListSummaryCard
+      isDeleting={isDeleting}
+      list={list}
+      onDelete={onDelete}
+      onOpen={onOpen}
+      variant="active"
+    />
   );
 };
