+++ client/src/pages/my-lists/hooks.ts
@@ -1,24 +1,14 @@
 import { useState } from "react";
-import { AuthenticatedNavigation, AuthenticatedStackRouteNames } from "@routes/types";
 import { getHttpErrorMessage } from "@services/http";
-import {
-  useCreateShoppingList,
-  useRemoveShoppingList,
-  useShoppingLists,
-} from "@hooks/use-shopping-lists";
-import { useMercadinNavigation } from "@hooks/use-navigation";
-import { ShoppingList } from "@services/shopping-lists/types";
+import { useCreateShoppingList, useShoppingLists } from "@hooks/use-shopping-lists";
+import { useShoppingListActions } from "@hooks/use-shopping-list-actions";
 
 export const useMyLists = () => {
-  const navigation = useMercadinNavigation<AuthenticatedNavigation>();
   const shoppingLists = useShoppingLists();
   const createShoppingList = useCreateShoppingList();
-  const removeShoppingList = useRemoveShoppingList();
+  const listActions = useShoppingListActions();
   const [name, setName] = useState("");
   const [submitError, setSubmitError] = useState<string | null>(null);
-  const [deleteError, setDeleteError] = useState<string | null>(null);
-  const [listToDelete, setListToDelete] = useState<ShoppingList | null>(null);
-  const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false);
 
   const createList = async () => {
     const trimmedName = name.trim();
@@ -37,47 +27,11 @@ export const useMyLists = () => {
     }
   };
 
-  const openList = (listId: string) => {
-    navigation.navigate(AuthenticatedStackRouteNames.SHOPPING_LIST_DETAILS, { listId });
-  };
-
-  const requestDeleteList = (list: ShoppingList) => {
-    setListToDelete(list);
-    setIsDeleteDialogOpen(true);
-  };
-
-  const deleteList = async () => {
-    if (!listToDelete) {
-      return;
-    }
-
-    setIsDeleteDialogOpen(false);
-    setDeleteError(null);
-
-    try {
-      await removeShoppingList.mutateAsync(listToDelete.id);
-      setListToDelete(null);
-    } catch (error) {
-      setDeleteError(await getHttpErrorMessage(error, "Não foi possível excluir a lista."));
-    }
-  };
-
-  const confirmDeleteList = () => {
-    void deleteList();
-  };
-
   return {
-    confirmDeleteList,
+    ...listActions,
     createList,
     createShoppingList,
-    deleteError,
-    isDeleteDialogOpen,
-    listToDelete,
     name,
-    openList,
-    requestDeleteList,
-    removeShoppingList,
-    setIsDeleteDialogOpen,
     setName,
     shoppingLists,
     submitError,
